TechWebLinks: refactors file location
